requirements.txt
beautifulsoup4==4.12.0
branca==0.6.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
folium==0.14.0
gpxpy==1.5.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.5.0
psycopg2-binary==2.9.5
requests==2.28.2
soupsieve==2.4
SQLAlchemy==2.0.8
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
